requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
contourpy==1.1.1
cycler==0.12.1
dnspython==2.6.1
et_xmlfile==2.0.0
fonttools==4.57.0
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
openpyxl==3.1.5
packaging==25.0
pandas==2.0.3
pillow==10.4.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pymongo==4.10.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
zipp==3.20.2
